Refactor backend routes by extracting controllers
