+++ server/auth/oauth2.ts
@@ -23,12 +23,17 @@ if (
         clientSecret: OAUTH2_CLIENT_SECRET,
         callbackURL: OAUTH2_CALLBACK_URL,
       },
-      (_accessToken: string, _refreshToken: string, _profile: any, cb: unknown) => {
+      (_accessToken: string, _refreshToken: string, _profile: any, cb: any) => {
         // OAuth2 user lookup is not yet implemented.
         // Do NOT replace this with a hardcoded identity — every OAuth2 user would
         // share the same account and see all other users' patient records.
         // Implement a real DB lookup (e.g. by profile.emails[0].value) before enabling.
         return cb(new Error("OAuth2 authentication is not yet configured for this application."));
+      }
+    )
+  );
+}
+
 import { Router, type Request, type Response } from "express";
 import crypto from "crypto";
 import bcrypt from "bcrypt";
